requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
pywebpush==2.0.0
manifest-generator==1.0.0
requests==2.31.0
